Do not return PrettyNames for imptofu results
When there was no Keybase user for given imptofu query, we were
returning back the query string as PrettyName, so it looked nice in the
result list. But these results are not formatted to the result list in
GUI anymore, so that benefit is no more. Removing this simplifies
UserBubble tooltip formatting for TeamBox.